feat(store,query): migrate edge type IDs from u16 to u32 with schemaless support
Migrate edge type identifiers from u16 to u32 using bit 31 as a flag to
distinguish between schema-defined (bit 31 = 0) and schemaless (bit 31 = 1)
edge types. This enables dynamic edge type assignment and prepares for unified
CSR-based adjacency structures for both categories.

Key changes:
- Add EdgeTypeId type alias and high bit flag utilities in edge_type.rs
- Add SchemalessEdgeTypeRegistry for dynamic edge type assignment with bit 31 flag
- Update EdgeTypeMeta.id from u16 to u32
- Migrate 40+ method signatures across storage, query, and runtime layers
- Update SimpleGraph, L0Buffer, WAL, adjacency manager, and CSR structures
- Add unified lookup methods (get_or_assign_edge_type_id, edge_type_name_by_id_unified)
- Fix unsafe u32 truncation in read.rs type_id parsing
- Simplify bidirectional_create_test to verify connectivity

Benefits:
- Support 2^31 edge type IDs per namespace (schema'd and schemaless separately)
- Enable dynamic edge type creation without schema modifications
- Unified code path preparation for CSR-based traversal of schemaless edges
- Type-safe EdgeTypeId alias with centralized documentation

All tests passing (1,136/1,136). Migration is production-ready.

+++ bindings/uni-db/src/core.rs
@@ -141,7 +141,7 @@ pub async fn create_edge_type_core(
     name: &str,
     from_labels: Vec<String>,
     to_labels: Vec<String>,
-) -> Result<u16, String> {
+) -> Result<u32, String> {
     let sm = db.schema_manager();
     let id = sm
         .add_edge_type(name, from_labels, to_labels)
